requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.7
python-dotenv==1.0.0
requests==2.32.4
PyJWT==2.10.1
psycopg[binary]==3.2.13
